agent-runtime/requirements.txt
fastapi>=0.109.0,<1.0
uvicorn>=0.27.0,<1.0
pyautogui>=0.9.54
pillow>=10.4.0,<12.0
mss>=9.0.1
pynput>=1.7.6
psutil>=5.9.7
httpx>=0.26.0,<1.0
websockets>=12.0
openai>=1.12.0,<2.0
anthropic>=0.18.1,<1.0
google-generativeai>=0.4.0,<1.0
pydantic>=2.6.0,<3.0
python-dotenv>=1.0.1
aiofiles>=23.2.1
loguru>=0.7.2
pyperclip>=1.8.2
numpy>=1.26.0,<3.0
comtypes>=1.4.0;sys_platform=="win32"
cryptography>=42.0.0,<44.0
